requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
pgvector==0.3.6
pymupdf==1.25.3
openai==1.59.9
anthropic==0.42.0
python-dotenv==1.0.1
